Build a local copy of the Afrobarometer database

Automates the downloading and merging of the Afrobarometer data. Fetches any missing, publicly available files, creates a master MonetDB database in options(afrobarometer.data), and merges available spatial data with the questionnaires for each round.

afrb_build(rounds = 1:6, overwrite_db = FALSE)

Arguments

rounds
A sequence of integers indicating which rounds to build. As of February 2017, there are six rounds. Default is seq(1, 6).

overwrite_db
A boolean value indicating whether or not the SQLite database in the local Afrobarometer data directory should be overwritten when afrb_build is run. Default is FALSE.

Value

Builds the database in options("afrobarometer.data"), downloading files as needed into subdirectories and merging the resulting data in the SQLite database file options("afrobarometer.sqlite").

Details

Meant to be run between afrb_dir and afrb_round.

See also

afrb_dir

Examples

not_run({
  library(afrobarometer)
  afrb_dir(path = "~/foo")
  # download first 3 rounds but do not overwrite the existing database.
  # This will fail if the database is already populated with merged tables.
  afrb_build(rounds = 3:4)
  # download all rounds and overwrite the existing database.
  afrb_build(rounds = 3:4, overwrite_db = TRUE)
})