Set the Afrobarometer data directory

Customize where the data is cached. Must be run explicitly by the user.

afrb_dir(path = tempdir())

Arguments

path
Character scalar containing the file path to set. Run afrb_dir() without any paramters to use base::tempdir().

Value

Sets the afrobarometer data directory and database file invisibly.

Details

Sets the options afrobarometer.data and afrobarometer.sqlite, the paths of the local Afrobarometer data cache and SQLite database respectively. The questionnaire data will be automatically downloaded into the subdirectory questionnaires. The user is expected to manually add spatial data CSV files in the format locations/Location_Rk.csv where locations is a subdirectory of the Afrobrometer data folder, afrobarometer.data. afrb_dir() automatically creates locations, so the user simply has to place the spatial data CSVs in the locations subdirectory after running afrb_dir().

Examples

library(afrobarometer)
afrb_dir("~/foo")

#> [afrb] Setting options(afrobarometer.data) to /Users/karencottriel/foo

#> [afrb] Spatial data should be placed in the `locations` subdirectory of the Afrobarometer data directory. For example, if you have spatial data for Rounds 3 and 4, you should place the spatial data as follows:

#> [afrb]

#> [afrb] /Users/karencottriel/foo/locations/Locations_R3.csv

#> [afrb] /Users/karencottriel/foo/locations/Locations_R4.csv

#> [afrb]

#> [afrb] To change the data directory to file path x, use afrobarometer::afrb_dir(x)

#> [afrb] Creating directories

#> [afrb] - /Users/karencottriel/foo (exists)

#> [afrb] - /Users/karencottriel/foo/questionnaires (exists)

#> [afrb] - /Users/karencottriel/foo/locations (exists)

#> [afrb] - /Users/karencottriel/foo/codebooks (exists)