afrobarometer: Load the Merged Afrobarometer Survey Data

The afrobarometer package automates downloading and merging the Afrobarometer Survey Data into the R workspace.

Afrobarometer functions

There are three exported functions of the afrobarometer package:

afrb_dir(path)

Sets the Afrobarometer data directory to path.

afrb_build(rounds, overwrite_db)

After running afrb_dir, the function afrb_build constructs the local database for the surveys specified in rounds. If overwrite_db is set to TRUE, the existing database file is overwritten.

afrb_round(round)

Extracts merged data for Round code from the loceal SQLite database into a tibble in R.

Open access data is downloaded if needed. The user is expected to manually set the optional restricted access data in the locations subdirectory of the Afrobarometer data directory, getOption("afrobarometer.data"), such as spatial locations of survey respondents.

Examples

not_run({
  library(afrobarometer)
  afrb_dir(path = "~/foo")
  # download first 3 rounds but do not overwrite the existing database file.
  # This will fail if the database file has tables.
  afrb_build(rounds = 3:4)
  # download all rounds (rounds 1 through 6) and overwrite the existing database file.
  afrb_build(rounds = 3:4, overwrite_db = TRUE)
  # load round 4 from the database into a data.frame
  afrb4 <- afrb_round(4)
})